+++ CarvedRock/CarvedRock.UITests/WindowsUWPLongTest.cs
@@ -99,5 +99,95 @@ namespace CarvedRock.UITests
             driver.CloseApp();
         }
 
+        [TestMethod]
+        public void AddNewItem()
+        {
+            var capabilities = new AppiumOptions();
+            capabilities.AddAdditionalCapability(MobileCapabilityType.App, "8b831c56-bc54-4a8b-af94-a448f80118e7_sezxftbtgh66j!App");
+            capabilities.AddAdditionalCapability(MobileCapabilityType.PlatformName, "Windows");
+            capabilities.AddAdditionalCapability(MobileCapabilityType.DeviceName, "WindowsPC");
+
+            var _appiumLocalService = new AppiumServiceBuilder().UsingAnyFreePort().Build();
+            _appiumLocalService.Start();
+            var driver = new WindowsDriver<WindowsElement>(_appiumLocalService, capabilities);
+
+            var el1 = driver.FindElementByAccessibilityId("Add");
+            el1.Click();
+
+            var elItemText = driver.FindElementByAccessibilityId("ItemText");
+            elItemText.Clear();
+            elItemText.SendKeys("This is a new Item");
+
+            var elItemDetail = driver.FindElementByAccessibilityId("ItemDescription");
+            elItemDetail.Clear();
+            elItemDetail.SendKeys("These are the details");
+
+            var elSave = driver.FindElementByAccessibilityId("Save");
+            elSave.Click();
+
+            //wait for progress bar to disapear
+            var wait = new DefaultWait<WindowsDriver<WindowsElement>>(driver)
+            {
+                Timeout = TimeSpan.FromSeconds(60),
+                PollingInterval = TimeSpan.FromMilliseconds(500)
+            };
+            wait.IgnoreExceptionTypes(typeof(NoSuchElementException));
+            wait.Until(d => d.FindElementByName("Second item"));
+
+            var listview = driver.FindElementByAccessibilityId("ItemsListView");
+
+            //now use wait to scroll untill we find item
+            wait = new DefaultWait<WindowsDriver<WindowsElement>>(driver)
+            {
+                Timeout = TimeSpan.FromSeconds(60),
+                PollingInterval = TimeSpan.FromMilliseconds(500)
+            };
+            wait.IgnoreExceptionTypes(typeof(NoSuchElementException));
+
+            var elementfound = wait.Until(d =>
+            {
+                var input = new PointerInputDevice(PointerKind.Touch);
+                ActionSequence FlickUp = new ActionSequence(input);
+                FlickUp.AddAction(input.CreatePointerMove(listview, 0, 0, TimeSpan.Zero));
+                FlickUp.AddAction(input.CreatePointerDown(MouseButton.Left));
+                FlickUp.AddAction(input.CreatePointerMove(listview, 0, -300, TimeSpan.FromMilliseconds(200)));
+                FlickUp.AddAction(input.CreatePointerUp(MouseButton.Left));
+                driver.PerformActions(new List<ActionSequence>() { FlickUp });
+
+                return d.FindElementByName("This is a new Item");
+            });
+
+            Assert.IsTrue(elementfound != null);
+
+            driver.CloseApp();
+
+        }
+        [TestMethod]
+        public void MasterDetail()
+        {
+            var capabilities = new AppiumOptions();
+            capabilities.AddAdditionalCapability(MobileCapabilityType.App, "8b831c56-bc54-4a8b-af94-a448f80118e7_sezxftbtgh66j!App");
+            capabilities.AddAdditionalCapability(MobileCapabilityType.PlatformName, "Windows");
+            capabilities.AddAdditionalCapability(MobileCapabilityType.DeviceName, "WindowsPC");
+
+            var _appiumLocalService = new AppiumServiceBuilder().UsingAnyFreePort().Build();
+            _appiumLocalService.Start();
+            var driver = new WindowsDriver<WindowsElement>(_appiumLocalService, capabilities);
+
+            // tap on second item
+            var el1 = driver.FindElementByName("Second item");
+
+            el1.Click();
+            var el2 = driver.FindElementByAccessibilityId("ItemText");
+            Assert.IsTrue(el2.Text == "Second item");
+
+            var backButton = driver.FindElementByAccessibilityId("Back");
+            backButton.Click();
+
+            var el3 = driver.FindElementByName("Fourth item");
+            Assert.IsTrue(el3 != null);
+
+            driver.CloseApp();
+        }
     }
 }
